refactor(META): Change shared utils to walker
This provides a better context for the class
